move rand back into remote module to avoid ts overflow
