+++ modules/common/utils/index.ts
@@ -10,4 +10,3 @@ export * from "./punctuate";
 export * from "./gameSlug";
 export * from "./gameCount";
 export * from "./whoWins";
-export * from "./random";
